# doc/api.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:3000/api/v3
info:
  title: WASAText
  description: Web And Software Architecture Exam Project
  version: 1.0.0
security:
  - bearerAuth: []
...
paths:
...
  /messages/{messageId}:
    parameters:
      - name: messageId
        in: path
        description: Message id
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: ["messages"]
      operationId: getMessage
      summary: Get message data
      description: Get message data
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: ["messages"]
      operationId: deleteMessage
      summary: Delete message
      description: Delete message
      responses:
        '204':
          description: No Content
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        description: User id
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: ["users"]
      summary: Get user data
      description: Get user data
      operationId: getUserById
      responses:
        '200':
          description: Successful
          content:
            application/json:
              example:
                type: object
              schema:
                $ref: "#/components/schemas/User"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: ["users"]
      summary: Delete user
      description: Delete user
      operationId: deleteUser
      responses:
        '204':
          description: No Content
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            description: Error
            type: object
            properties:
              error:
                description: Message
                type: string
  schemas:
    User:
      description: User
      type: object
      properties:
        userId:
          description: User id
          type: integer
          format: int64
        name:
          $ref: "#/components/schemas/Name"
        photoUrl:
          $ref: "#/components/schemas/PhotoUrl"
...
    Name:
      description: Generic name
      type: string
      pattern: '^.*$'
      minLength: 1
      maxLength: 1024
    PhotoUrl:
      description: Photo URL
      type: string
      format: url
      minLength: 1
      maxLength: 1024
    Message:
      description: Message
      type: object
      properties:
        messageId:
          description: Message id
          type: integer
          format: int64
        text:
          description: Message text
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 1024
        senderId:
          description: Sender id
          type: integer
          format: int64
        conversationId:
          description: Conversation id
          type: integer
          format: int64
        createdAt:
          description: Message timestamp
          type: string
          format: date-time
        isForwarded:
          description: Message is forwarded
          type: boolean
        commentTo:
          description: Commented message id
          type: integer
          format: int64
        attachmentUrl:
          description: Attachment URL
          type: string
          format: url
        mediaType:
          $ref: "#/components/schemas/MediaType"
...
    MediaType:
      description: Media type
      type: string
      enum: [image, video, audio, file]
...